Initial cellular automata simulation
Hi Guys,
Have a look at the following.
The sim currently only aggregates the position of the vehicles to the output matrix. It is simple to collect other state variables. I'll do that next.
To run the SIM, download the files into your working directory and run the 'RunCellAutomata' script.
Cheers - Mark